Revert "Try to fix array iteration"
This reverts commit b2cf0300ba4fb377eb0ff15def629fa8922b32dc.
